+++ R/utils.R
@@ -519,22 +519,53 @@ cramer4 <- function(object, imp) {
 ## }
 
 getColsFromPattern <- function(x, pattern) {
-  if (missing(pattern))
-    stop("Pattern must not be missing.")
-  if (nchar(pattern) != ncol(x))
-    stop("The pattern must be equal to the number of columns.")
-  pattern <- strsplit(pattern, "")[[1]]
-  if (!all(unique(pattern) %in% c("0", "1")))
-    stop("Pattern must be composed of '0' or '1' defining columns with or without 'NA's.")
-  return(pattern == "1")
+  if (missing(pattern)) {
+    stop(sQuote("pattern"), " must not be missing.")
+  }
+  if (!is.matrix(x)) {
+    stop(sQuote("x"), " must be a matrix.")
+  }
+  if (nchar(pattern) != ncol(x)) {
+    stop("The ", sQuote("pattern"), " must be equal to the number of columns.")
+  }
+  pattern <- strsplit(pattern, "")[[1L]]
+  if (!all(unique(pattern) %in% c("0", "1"))) {
+    stop(sQuote("pattern"), " must be composed of '0' or '1' defining columns",
+         " with or without 'NA's.")
+  }
+  pattern == "1"
 }
 
 getRowsFromPattern <- function(x, pattern) {
   cols <- getColsFromPattern(x, pattern)
-  x2 <- x[, cols]
-  apply(x2, 1, function(xx) !any(is.na(xx)))
+  x <- x[, cols, drop=FALSE]
+  rowSums(is.na(x)) == 0
 }
 
+.filterNA <- function(x, pNA=0L, pattern) {
+  if (!is.matrix(x)) {
+    stop(sQuote("x"), " must be a matrix.")
+  }
+  if (!is.numeric(pNA)) {
+    stop(sQuote("pNA"), " must be numeric.")
+  }
+  if (length(pNA) > 1) {
+    stop(sQuote("pNA"), " must be of length one.")
+  }
+
+  if (missing(pattern)) { ## using pNA
+    if (pNA > 1) {
+      pNA <- 1
+    }
+    if (pNA < 0) {
+      pNA <- 0
+    }
+    k <- rowSums(is.na(x)) / ncol(x)
+    k <= pNA
+  } else { ## using pattern
+    getRowsFromPattern(x, pattern)
+  }
+}
 
 nologging <- function(object, n = 1) {
   ## removes the last n entries from
@@ -1026,3 +1057,50 @@ countAndPrint <- function(x) {
   if (inherits(object, "OnDiskMSnExp")) msLevel(object)[1]
   else msLevel(object[[1]])
 }
+
+#' @title Define the type of mzR backend to use based on the file name
+#'
+#' @description Simple helper to define the mzR backend that should/can be used
+#'     to read the file.
+#'
+#' @param x \code{character(1)} representing the file name.
+#'
+#' @return A \code{character(1)} with the name of the backend (either
+#'     \code{"netCDF"}, \code{"Ramp"} or \code{"pwiz"}.
+#' 
+#' @author Johannes Rainer, Sebastian Gibb
+#'
+#' @noRd
+.mzRBackend <- function(x = character()) {
+    if (length(x) != 1)
+        stop("parameter 'x' has to be of length 1")
+    ## Use if/else conditions based on a suggestion from sgibb to avoid loops.
+    if (grepl("\\.mzml($|\\.)|\\.mzxml($|\\.)", x, ignore.case = TRUE)) {
+        return("pwiz")
+    } else if (grepl("\\.mzdata($|\\.)", x, ignore.case = TRUE)) {
+        return("Ramp")
+    } else if (grepl("\\.cdf($|\\.)|\\.nc($|\\.)", x, ignore.case = TRUE)) {
+        return("netCDF")
+    } else {
+        stop("Could not determine file type for ", x)
+    }
+}
+
+#' @title Open an MS file using the mzR package
+#'
+#' @description Opens an MS file using the mzR package determining the corrent
+#'     backend based on the file ending of the specified file.
+#'
+#' @param x \code{character(1)}: the file name.
+#'
+#' @return A file handle to the opened MS file.
+#'
+#' @author Johannes Rainer
+#' 
+#' @noRd
+.openMSfile <- function(x) {
+    if (missing(x) || length(x) != 1)
+        stop("parameter 'x' has to be of length 1")
+    mzR::openMSfile(x, backend = .mzRBackend(x))
+}
+