+++ NEWS.md
@@ -5,6 +5,7 @@
     <2017-05-11 Thu>.
 - Add `.filterNA` and rewrite `filterNA` for `matrix` and `MSnSet`
   <2017-05-11 Thu>.
+- Convert main MSnbase-demo vignette to Rmd/html <2017-05-27 Sat>
 
 ## Changes in version 2.3.2
 - Rewrite `utils.clean`. It now keeps just the zeros in the direct