+++ NEWS.md
@@ -6,6 +6,8 @@
 - Add `.filterNA` and rewrite `filterNA` for `matrix` and `MSnSet`
   <2017-05-11 Thu>.
 - Convert main MSnbase-demo vignette to Rmd/html <2017-05-27 Sat>
+- `naplot` gains a `reorderRows` and `reorderColumns` argument
+  <2017-06-05 Mon>.
 
 ## Changes in version 2.3.2
 - Rewrite `utils.clean`. It now keeps just the zeros in the direct